+++ crates/bevy_gltf/src/loader.rs
@@ -4,7 +4,7 @@ use bevy_asset::{
 };
 use bevy_core::Name;
 use bevy_ecs::world::World;
-use bevy_hierarchy::hierarchy::{BuildWorldChildren, WorldChildBuilder};
+use bevy_hierarchy::{BuildWorldChildren, WorldChildBuilder};
 use bevy_log::warn;
 use bevy_math::{Mat4, Vec3};
 use bevy_pbr::{
