+++ crates/bevy_hierarchy/Cargo.toml
@@ -15,7 +15,6 @@ bevy_ecs = { path = "../bevy_ecs", version = "0.6.0", features = ["bevy_reflect"
 bevy_math = { path = "../bevy_math", version = "0.6.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.6.0", features = ["bevy"] }
 bevy_utils = { path = "../bevy_utils", version = "0.6.0" }
-bevy_transform = {path = "../bevy_transform", version = "0.6.0"}
 
 # other
 smallvec = { version = "1.6", features = ["serde", "union", "const_generics"] }
