--- crates/bevy_hierarchy/src/hierarchy/mod.rs
@@ -1,8 +1,0 @@
-mod child_builder;
-#[allow(clippy::module_inception)]
-mod hierarchy;
-mod hierarchy_maintenance_system;
-
-pub use child_builder::*;
-pub use hierarchy::*;
-pub use hierarchy_maintenance_system::*;
