+++ crates/bevy_hierarchy/src/lib.rs
@@ -5,11 +5,18 @@
 //! from the [`Parent`] to its [`Children`].
 
 /// The basic components of the hierarchy
-pub mod components;
+mod components;
+pub use components::*;
+
 /// Establishing and updating the transform hierarchy
-pub mod hierarchy;
-/// Propagating transform changes down the transform hierarchy
-pub mod transform_propagate_system;
+mod hierarchy;
+pub use hierarchy::*;
+
+mod child_builder;
+pub use child_builder::*;
+
+mod systems;
+pub use systems::*;
 
 #[doc(hidden)]
 pub mod prelude {
@@ -18,18 +25,15 @@ pub mod prelude {
 }
 
 use bevy_app::prelude::*;
-use bevy_ecs::schedule::{ParallelSystemDescriptorCoercion, SystemLabel};
-use prelude::{parent_update_system, Children, Parent, PreviousParent};
+use bevy_ecs::prelude::*;
 
 /// The base plugin for handling [`Parent`] and [`Children`] components
 #[derive(Default)]
 pub struct HierarchyPlugin;
 
-/// Label enum for the types of systems relating to transform
+/// Label enum for the systems relating to hierarchy upkeep
 #[derive(Debug, Hash, PartialEq, Eq, Clone, SystemLabel)]
 pub enum HierarchySystem {
-    /// Propagates changes in transform to childrens' [`GlobalTransform`](bevy_transform::components::GlobalTransform)
-    TransformPropagate,
     /// Updates [`Parent`] when changes in the hierarchy occur
     ParentUpdate,
 }
@@ -39,26 +43,13 @@ impl Plugin for HierarchyPlugin {
         app.register_type::<Children>()
             .register_type::<Parent>()
             .register_type::<PreviousParent>()
-            // add transform systems to startup so the first update is "correct"
             .add_startup_system_to_stage(
                 StartupStage::PostStartup,
                 parent_update_system.label(HierarchySystem::ParentUpdate),
             )
-            .add_startup_system_to_stage(
-                StartupStage::PostStartup,
-                transform_propagate_system::transform_propagate_system
-                    .label(HierarchySystem::TransformPropagate)
-                    .after(HierarchySystem::ParentUpdate),
-            )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
                 parent_update_system.label(HierarchySystem::ParentUpdate),
-            )
-            .add_system_to_stage(
-                CoreStage::PostUpdate,
-                transform_propagate_system::transform_propagate_system
-                    .label(HierarchySystem::TransformPropagate)
-                    .after(HierarchySystem::ParentUpdate),
             );
     }
 }
