+++ crates/bevy_pbr/Cargo.toml
@@ -18,7 +18,6 @@ bevy_asset = { path = "../bevy_asset", version = "0.6.0" }
 bevy_core = { path = "../bevy_core", version = "0.6.0" }
 bevy_core_pipeline = { path = "../bevy_core_pipeline", version = "0.6.0" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.6.0" }
-bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.6.0" }
 bevy_math = { path = "../bevy_math", version = "0.6.0" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.6.0", features = ["bevy"] }
 bevy_render = { path = "../bevy_render", version = "0.6.0" }
