+++ crates/bevy_scene/src/scene_spawner.rs
@@ -7,7 +7,7 @@ use bevy_ecs::{
     system::Command,
     world::{Mut, World},
 };
-use bevy_hierarchy::{hierarchy::AddChild, prelude::Parent};
+use bevy_hierarchy::{AddChild, Parent};
 use bevy_reflect::TypeRegistryArc;
 use bevy_utils::{tracing::error, HashMap};
 use thiserror::Error;
