+++ crates/bevy_ui/src/flex/mod.rs
@@ -7,7 +7,7 @@ use bevy_ecs::{
     query::{Changed, FilterFetch, With, Without, WorldQuery},
     system::{Query, Res, ResMut},
 };
-use bevy_hierarchy::components::{Children, Parent};
+use bevy_hierarchy::{Children, Parent};
 use bevy_log::warn;
 use bevy_math::Vec2;
 use bevy_transform::components::Transform;
