+++ crates/bevy_ui/src/lib.rs
@@ -27,9 +27,9 @@ pub mod prelude {
 
 use bevy_app::prelude::*;
 use bevy_ecs::schedule::{ParallelSystemDescriptorCoercion, SystemLabel};
-use bevy_hierarchy::HierarchySystem;
 use bevy_input::InputSystem;
 use bevy_math::{Rect, Size};
+use bevy_transform::TransformSystem;
 use update::{ui_z_system, update_clipping_system};
 
 use crate::prelude::CameraUi;
@@ -93,17 +93,17 @@ impl Plugin for UiPlugin {
                 CoreStage::PostUpdate,
                 flex_node_system
                     .label(UiSystem::Flex)
-                    .before(HierarchySystem::TransformPropagate),
+                    .before(TransformSystem::TransformPropagate),
             )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
                 ui_z_system
                     .after(UiSystem::Flex)
-                    .before(HierarchySystem::TransformPropagate),
+                    .before(TransformSystem::TransformPropagate),
             )
             .add_system_to_stage(
                 CoreStage::PostUpdate,
-                update_clipping_system.after(HierarchySystem::TransformPropagate),
+                update_clipping_system.after(TransformSystem::TransformPropagate),
             );
 
         crate::render::build_ui_render(app);
